requirements.txt
# Core Framework
streamlit==1.38.0

# Environment Variables
python-dotenv==1.0.0

# AI Assistant
openai==1.12.0

# Web APIs
requests==2.31.0

# Music Processing
music21==9.1.0
basic-pitch==0.3.2
mido==1.3.2
pretty-midi==0.2.10

# Audio Processing
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.4

# File Handling
python-multipart==0.0.9

# Utilities
pandas==2.2.0
matplotlib==3.8.2

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3

# YouTube Downloader
yt-dlp==2023.12.30

# Optional: For better performance
scipy==1.12.0
